Refactored polaris code for sep-6
