+++ example/server/settings.py
@@ -42,6 +42,8 @@ INSTALLED_APPS = [
     "sslserver",
 ]
 
+ACTIVE_SEPS = ["sep-1", "sep-6", "sep-10", "sep-24"]
+
 MIDDLEWARE = [
     "django.middleware.security.SecurityMiddleware",
     "corsheaders.middleware.CorsMiddleware",
