--- polaris/polaris/fee/urls.py
@@ -1,5 +1,0 @@
-"""This module defines the URL patterns for the `/fee` endpoint."""
-from django.urls import path
-from polaris.fee.views import fee
-
-urlpatterns = [path("", fee)]
