+++ polaris/polaris/management/commands/poll_pending_deposits.py
@@ -2,10 +2,10 @@ import time
 
 from django.core.management import BaseCommand, CommandError
 
-from polaris.deposit.utils import create_stellar_deposit
+from polaris.utils import create_stellar_deposit
 from polaris.integrations import registered_deposit_integration as rdi
 from polaris.models import Transaction
-from polaris.helpers import Logger
+from polaris.utils import Logger
 
 logger = Logger(__name__)
