+++ polaris/polaris/sep1/urls.py
@@ -1,5 +1,5 @@
 """This module defines the URL patterns for the `/.well-known/stellar.toml` endpoint."""
 from django.urls import path
-from polaris.stellartoml.views import generate_toml
+from polaris.sep1.views import generate_toml
 
-urlpatterns = [path("/stellar.toml", generate_toml)]
+urlpatterns = [path("stellar.toml", generate_toml)]
