+++ polaris/polaris/sep1/views.py
@@ -16,8 +16,8 @@ from polaris.integrations import registered_toml_func
 def generate_toml(request):
     """Generate the TOML file."""
     toml_dict = {
-        "TRANSFER_SERVER": settings.HOST_URL,
-        "TRANSFER_SERVER_SEP0024": settings.HOST_URL,
+        "TRANSFER_SERVER": os.path.join(settings.HOST_URL, "sep6"),
+        "TRANSFER_SERVER_SEP0024": os.path.join(settings.HOST_URL, "sep24"),
         "WEB_AUTH_ENDPOINT": os.path.join(settings.HOST_URL, "auth"),
         "ACCOUNTS": [
             asset["DISTRIBUTION_ACCOUNT_ADDRESS"] for asset in settings.ASSETS.values()
