+++ polaris/polaris/sep24/info.py
@@ -5,26 +5,31 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 
 from polaris.models import Asset
-from polaris.integrations import registered_fee_func, calculate_fee
+from polaris.integrations import (
+    registered_fee_func,
+    calculate_fee,
+)
 
 
 def _get_asset_info(asset: Asset, op_type: str) -> Dict:
-    if getattr(asset, f"{op_type}_enabled"):
-        asset_info = {
-            "enabled": True,
-            "min_amount": getattr(asset, f"{op_type}_min_amount"),
-            "max_amount": getattr(asset, f"{op_type}_max_amount"),
-        }
-        if registered_fee_func == calculate_fee:
-            # the anchor has not replaced the default fee function
-            # so `fee_fixed` and `fee_percent` are still relevant.
-            asset_info.update(
-                fee_fixed=getattr(asset, f"{op_type}_fee_fixed"),
-                fee_percent=getattr(asset, f"{op_type}_fee_percent"),
-            )
-        return asset_info
+    if not getattr(asset, f"{op_type}_enabled"):
+        return {"enabled": False}
 
-    return {"enabled": False}
+    asset_info = {
+        "enabled": True,
+        "authentication_required": True,
+        "min_amount": getattr(asset, f"{op_type}_min_amount"),
+        "max_amount": getattr(asset, f"{op_type}_max_amount"),
+    }
+    if registered_fee_func == calculate_fee:
+        # the anchor has not replaced the default fee function
+        # so `fee_fixed` and `fee_percent` are still relevant.
+        asset_info.update(
+            fee_fixed=getattr(asset, f"{op_type}_fee_fixed"),
+            fee_percent=getattr(asset, f"{op_type}_fee_percent"),
+        )
+
+    return asset_info
 
 
 @api_view()
@@ -42,7 +47,7 @@ def info(request):
         "transaction": {"enabled": True},
     }
 
-    for asset in Asset.objects.all():
+    for asset in Asset.objects.filter(sep24_enabled=True):
         info_data["deposit"][asset.code] = _get_asset_info(asset, "deposit")
         info_data["withdraw"][asset.code] = _get_asset_info(asset, "withdrawal")
