+++ polaris/polaris/tests/fee_test.py
@@ -10,12 +10,15 @@ from polaris import settings
 from polaris.tests.helpers import mock_check_auth_success
 
 
+endpoint = "/sep24/fee"
+
+
 @pytest.mark.django_db
-@patch("polaris.helpers.check_auth", side_effect=mock_check_auth_success)
+@patch("polaris.sep10.utils.check_auth", side_effect=mock_check_auth_success)
 def test_fee_no_params(mock_check, client):
     """Fails with no params provided."""
     del mock_check
-    response = client.get(f"/fee", follow=True)
+    response = client.get(endpoint, follow=True)
     content = json.loads(response.content)
 
     assert response.status_code == 400
@@ -23,11 +26,11 @@ def test_fee_no_params(mock_check, client):
 
 
 @pytest.mark.django_db
-@patch("polaris.helpers.check_auth", side_effect=mock_check_auth_success)
+@patch("polaris.sep10.utils.check_auth", side_effect=mock_check_auth_success)
 def test_fee_wrong_asset_code(mock_check, client):
     """Fails with an invalid `asset_code`."""
     del mock_check
-    response = client.get(f"/fee?asset_code=NADA", follow=True)
+    response = client.get(f"{endpoint}?asset_code=NADA", follow=True)
     content = json.loads(response.content)
 
     assert response.status_code == 400
@@ -35,12 +38,12 @@ def test_fee_wrong_asset_code(mock_check, client):
 
 
 @pytest.mark.django_db
-@patch("polaris.helpers.check_auth", side_effect=mock_check_auth_success)
+@patch("polaris.sep10.utils.check_auth", side_effect=mock_check_auth_success)
 def test_fee_no_operation(mock_check, client, usd_asset_factory):
     """Fails with no `operation` provided."""
     del mock_check
     usd_asset_factory()
-    response = client.get(f"/fee?asset_code=USD&amount=100", follow=True)
+    response = client.get(f"{endpoint}?asset_code=USD&amount=100", follow=True)
     content = json.loads(response.content)
 
     assert response.status_code == 400
@@ -48,13 +51,13 @@ def test_fee_no_operation(mock_check, client, usd_asset_factory):
 
 
 @pytest.mark.django_db
-@patch("polaris.helpers.check_auth", side_effect=mock_check_auth_success)
+@patch("polaris.sep10.utils.check_auth", side_effect=mock_check_auth_success)
 def test_fee_invalid_operation(mock_check, client, usd_asset_factory):
     """Fails with an invalid `operation` provided."""
     del mock_check
     usd_asset_factory()
     response = client.get(
-        f"/fee?asset_code=USD&amount=100&operation=generate", follow=True
+        f"{endpoint}?asset_code=USD&amount=100&operation=generate", follow=True
     )
     content = json.loads(response.content)
 
@@ -63,12 +66,12 @@ def test_fee_invalid_operation(mock_check, client, usd_asset_factory):
 
 
 @pytest.mark.django_db
-@patch("polaris.helpers.check_auth", side_effect=mock_check_auth_success)
+@patch("polaris.sep10.utils.check_auth", side_effect=mock_check_auth_success)
 def test_fee_no_amount(mock_check, client, usd_asset_factory):
     """Fails with no amount provided."""
     del mock_check
     usd_asset_factory()
-    response = client.get(f"/fee?asset_code=USD&operation=withdraw", follow=True)
+    response = client.get(f"{endpoint}?asset_code=USD&operation=withdraw", follow=True)
     content = json.loads(response.content)
 
     assert response.status_code == 400
@@ -76,13 +79,13 @@ def test_fee_no_amount(mock_check, client, usd_asset_factory):
 
 
 @pytest.mark.django_db
-@patch("polaris.helpers.check_auth", side_effect=mock_check_auth_success)
+@patch("polaris.sep10.utils.check_auth", side_effect=mock_check_auth_success)
 def test_fee_invalid_amount(mock_check, client, usd_asset_factory):
     """Fails with a non-float amount provided."""
     del mock_check
     usd_asset_factory()
     response = client.get(
-        f"/fee?asset_code=USD&operation=withdraw&amount=TEXT", follow=True
+        f"{endpoint}?asset_code=USD&operation=withdraw&amount=TEXT", follow=True
     )
     content = json.loads(response.content)
 
@@ -91,14 +94,14 @@ def test_fee_invalid_amount(mock_check, client, usd_asset_factory):
 
 
 @pytest.mark.django_db
-@patch("polaris.helpers.check_auth", side_effect=mock_check_auth_success)
+@patch("polaris.sep10.utils.check_auth", side_effect=mock_check_auth_success)
 def test_fee_withdraw_disabled(mock_check, client, eth_asset_factory):
     """Fails if the withdraw `operation` is not enabled for the `asset_code`."""
     del mock_check
     eth_asset_factory()
 
     response = client.get(
-        f"/fee?asset_code=ETH&operation=withdraw&amount=100.0", follow=True
+        f"{endpoint}?asset_code=ETH&operation=withdraw&amount=100.0", follow=True
     )
     content = json.loads(response.content)
 
@@ -107,7 +110,7 @@ def test_fee_withdraw_disabled(mock_check, client, eth_asset_factory):
 
 
 @pytest.mark.django_db
-@patch("polaris.helpers.check_auth", side_effect=mock_check_auth_success)
+@patch("polaris.sep10.utils.check_auth", side_effect=mock_check_auth_success)
 def test_fee_deposit_disabled(mock_check, client, eth_asset_factory):
     """Fails if the deposit `operation` is not enabled for `asset_code`."""
     del mock_check
@@ -116,7 +119,7 @@ def test_fee_deposit_disabled(mock_check, client, eth_asset_factory):
     asset.save()
 
     response = client.get(
-        f"/fee?asset_code=ETH&operation=deposit&amount=100.0", follow=True
+        f"{endpoint}?asset_code=ETH&operation=deposit&amount=100.0", follow=True
     )
     content = json.loads(response.content)
 
@@ -125,14 +128,14 @@ def test_fee_deposit_disabled(mock_check, client, eth_asset_factory):
 
 
 @pytest.mark.django_db
-@patch("polaris.helpers.check_auth", side_effect=mock_check_auth_success)
+@patch("polaris.sep10.utils.check_auth", side_effect=mock_check_auth_success)
 def test_fee_valid_deposit(mock_check, client, usd_asset_factory):
     """Succeeds for a valid deposit."""
     del mock_check
     usd_asset_factory()
 
     response = client.get(
-        f"/fee?asset_code=USD&operation=deposit&amount=200.0", follow=True
+        f"{endpoint}?asset_code=USD&operation=deposit&amount=200.0", follow=True
     )
     content = json.loads(response.content)
 
@@ -142,14 +145,14 @@ def test_fee_valid_deposit(mock_check, client, usd_asset_factory):
 
 # Fixed: 5.0 Percent = 1
 @pytest.mark.django_db
-@patch("polaris.helpers.check_auth", side_effect=mock_check_auth_success)
+@patch("polaris.sep10.utils.check_auth", side_effect=mock_check_auth_success)
 def test_fee_valid_withdrawal(mock_check, client, usd_asset_factory):
     """Succeeds for a valid withdrawal."""
     del mock_check
     usd_asset_factory()
 
     response = client.get(
-        f"/fee?asset_code=USD&operation=withdraw&amount=100.0", follow=True
+        f"{endpoint}?asset_code=USD&operation=withdraw&amount=100.0", follow=True
     )
     content = json.loads(response.content)
 
@@ -160,12 +163,14 @@ def test_fee_valid_withdrawal(mock_check, client, usd_asset_factory):
 @pytest.mark.django_db
 def test_fee_authenticated_success(client, usd_asset_factory):
     """Succeeds for a valid fee, with successful authentication."""
+    from polaris.tests.auth_test import endpoint as auth_endpoint
+
     usd_asset_factory()
     client_address = "GDKFNRUATPH4BSZGVFDRBIGZ5QAFILVFRIRYNSQ4UO7V2ZQAPRNL73RI"
     client_seed = "SDKWSBERDHP3SXW5A3LXSI7FWMMO5H7HG33KNYBKWH2HYOXJG2DXQHQY"
 
     # SEP 10.
-    response = client.get(f"/auth?account={client_address}", follow=True)
+    response = client.get(f"{auth_endpoint}?account={client_address}", follow=True)
     content = json.loads(response.content)
     envelope_xdr = content["transaction"]
     envelope_object = TransactionEnvelope.from_xdr(
@@ -176,7 +181,7 @@ def test_fee_authenticated_success(client, usd_asset_factory):
     client_signed_envelope_xdr = envelope_object.to_xdr()
 
     response = client.post(
-        "/auth",
+        auth_endpoint,
         data={"transaction": client_signed_envelope_xdr},
         content_type="application/json",
     )
@@ -189,7 +194,9 @@ def test_fee_authenticated_success(client, usd_asset_factory):
     header = {"HTTP_AUTHORIZATION": f"Bearer {encoded_jwt}"}
 
     response = client.get(
-        f"/fee?asset_code=USD&operation=withdraw&amount=100.0", follow=True, **header
+        f"{endpoint}?asset_code=USD&operation=withdraw&amount=100.0",
+        follow=True,
+        **header,
     )
     content = json.loads(response.content)
 
@@ -202,7 +209,7 @@ def test_fee_no_jwt(client, usd_asset_factory):
     """`GET /fee` fails if a required JWT is not provided."""
     usd_asset_factory()
     response = client.get(
-        f"/fee?asset_code=USD&operation=withdraw&amount=100.0", follow=True
+        f"{endpoint}?asset_code=USD&operation=withdraw&amount=100.0", follow=True
     )
     content = json.loads(response.content)
     assert response.status_code == 403
