+++ polaris/settings.py
@@ -43,6 +43,8 @@ third_party_apps.append("polaris")
 
 INSTALLED_APPS = django_apps + third_party_apps
 
+ACTIVE_SEPS = ["sep-1", "sep-10", "sep-24"]
+
 # Modules to add to parent project's MIDDLEWARE
 MIDDLEWARE = [
     "corsheaders.middleware.CorsMiddleware",
